Split onContainerChanged to pre and post callbacks for more granular use.
