+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Tests/Runtime/IUnityDependentExtensionsTests/Mocks/MockMonoDependent.cs
@@ -1,7 +1,7 @@
 ﻿using Chopsticks.Dependencies;
 using MonoContainerTests.Mocks;
 
-namespace IUnityDependentExtensionsTests
+namespace IUnityDependentExtensionsTests.Mocks
 {
     public class MockMonoDependent : 
         BaseMonoDependent<MockDependencyContainer, MockMonoContainerService>
